feat: WIP plugin system and first auth plugin POC
