--- auth/constants.go
@@ -1,17 +1,0 @@
-package auth
-
-const (
-	FieldID        = "id"
-	FieldCreatedAt = "created_at"
-	FieldUpdatedAt = "updated_at"
-)
-
-const (
-	UsersTable = "users"
-)
-
-const (
-	DefaultPageSize = 20
-	MaxPageSize     = 1000
-	MinPageSize     = 1
-)
