+++ plugin/registry.go
@@ -6,13 +6,11 @@ import (
 	"github.com/gofiber/fiber/v2"
 )
 
-// PluginRegistry manages all registered plugins
 type PluginRegistry struct {
 	globalPlugins []GlobalPlugin
 	routePlugins  map[string]RoutePlugin
 }
 
-// NewPluginRegistry creates a new plugin registry
 func NewPluginRegistry() *PluginRegistry {
 	return &PluginRegistry{
 		globalPlugins: make([]GlobalPlugin, 0),
@@ -20,17 +18,14 @@ func NewPluginRegistry() *PluginRegistry {
 	}
 }
 
-// RegisterGlobal adds a global plugin to the registry
 func (r *PluginRegistry) RegisterGlobal(plugin GlobalPlugin) {
 	r.globalPlugins = append(r.globalPlugins, plugin)
 }
 
-// RegisterRoute adds a route-level plugin to the registry
 func (r *PluginRegistry) RegisterRoute(plugin RoutePlugin) {
 	r.routePlugins[plugin.Name()] = plugin
 }
 
-// ApplyGlobal applies all registered global plugins to the Fiber app in order
 func (r *PluginRegistry) ApplyGlobal(app *fiber.App) error {
 	for _, plugin := range r.globalPlugins {
 		handler := plugin.Handler()
@@ -42,18 +37,15 @@ func (r *PluginRegistry) ApplyGlobal(app *fiber.App) error {
 	return nil
 }
 
-// GetRoutePlugin returns a route plugin by name
 func (r *PluginRegistry) GetRoutePlugin(name string) (RoutePlugin, bool) {
 	plugin, exists := r.routePlugins[name]
 	return plugin, exists
 }
 
-// GetRoutePlugins returns all registered route-level plugins
 func (r *PluginRegistry) GetRoutePlugins() map[string]RoutePlugin {
 	return r.routePlugins
 }
 
-// GetGlobalPlugins returns all registered global plugins
 func (r *PluginRegistry) GetGlobalPlugins() []GlobalPlugin {
 	return r.globalPlugins
 }
