+++ plugins/ratelimit/ratelimit.go
@@ -1,0 +1,61 @@
+package ratelimit
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/fiber/v2/middleware/limiter"
+	"github.com/nicolasbonnici/gorest/plugin"
+	"github.com/nicolasbonnici/gorest/pluginloader"
+)
+
+type RateLimitPlugin struct {
+	requestsPerSecond int
+	burst             int
+}
+
+func init() {
+	pluginloader.RegisterGlobalPluginFactory("ratelimit", func() plugin.GlobalPlugin {
+		return &RateLimitPlugin{
+			requestsPerSecond: 100,
+			burst:             200,
+		}
+	})
+}
+
+func (p *RateLimitPlugin) Name() string {
+	return "ratelimit"
+}
+
+func (p *RateLimitPlugin) Initialize(config map[string]interface{}) error {
+	if rps, ok := config["requests_per_second"].(int); ok {
+		p.requestsPerSecond = rps
+	} else if rps, ok := config["requests_per_second"].(float64); ok {
+		p.requestsPerSecond = int(rps)
+	}
+
+	if burst, ok := config["burst"].(int); ok {
+		p.burst = burst
+	} else if burst, ok := config["burst"].(float64); ok {
+		p.burst = int(burst)
+	}
+
+	return nil
+}
+
+func (p *RateLimitPlugin) Handler() fiber.Handler {
+	return limiter.New(limiter.Config{
+		Max:        p.requestsPerSecond,
+		Expiration: 1 * time.Second,
+		Storage:    nil,
+		KeyGenerator: func(c *fiber.Ctx) string {
+			return c.IP()
+		},
+		LimitReached: func(c *fiber.Ctx) error {
+			return c.Status(fiber.StatusTooManyRequests).JSON(fiber.Map{
+				"error": fmt.Sprintf("Rate limit exceeded: %d requests per second", p.requestsPerSecond),
+			})
+		},
+	})
+}
